+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **1.1.11-dev**
 
+  - fixed remote validator method (#130)
   - added `change` auto binded event for select inputs to have same behavior w/ fields
   - fixed bug on radio buttons and required constraint
   - added afterDate and beforeDate validators in parsley.extend